# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Contabilidade Pessoal API",
    "version": "0.0.1",
    "description": "API para gerenciamento e consulta de contas contábeis.\n\nObservações:\n- O parâmetro de consulta `view` controla a forma de retorno: `flat` (padrão) ou `tree`.\n- Em `flat`, os itens seguem a estrutura de `ContaFlatDto`.\n- Em `tree`, os itens seguem a estrutura de `ContaTreeDto` (recursiva via `inferiores`)."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor local"
    }
  ],
  "tags": [
    {
      "name": "Contas",
      "description": "Endpoints de consulta de contas"
    }
  ],
  "paths": {
    "/contas": {
      "get": {
        "tags": [
          "Contas"
        ],
        "summary": "Listar contas",
        "operationId": "listarContas",
        "parameters": [
          {
            "in": "query",
            "name": "view",
            "description": "Define a forma de visualização dos dados",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "flat",
                "tree"
              ],
              "default": "flat"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de contas",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ContaFlatDto"
                      }
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ContaTreeDto"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contas"
        ],
        "summary": "Criar nova conta",
        "operationId": "criarConta",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContaNewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conta criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaFlatDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          }
        }
      }
    },
    "/contas/{id}": {
      "get": {
        "tags": [
          "Contas"
        ],
        "summary": "Obter conta por ID",
        "operationId": "obterContaPorId",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Identificador da conta",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "view",
            "in": "query",
            "description": "Define a forma de visualização do retorno",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "flat",
                "tree"
              ],
              "default": "flat"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conta encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ContaFlatDto"
                    },
                    {
                      "$ref": "#/components/schemas/ContaTreeDto"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Conta não encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "Contas"
        ],
        "summary": "Atualizar conta por ID",
        "operationId": "atualizarConta",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Identificador da conta",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContaFlatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conta atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContaFlatDto"
                }
              }
            }
          },
          "404": {
            "description": "Conta não encontrada"
          }
        }
      },
      "delete": {
        "tags": [
          "Contas"
        ],
        "summary": "Deletar conta por ID",
        "operationId": "deletarConta",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Identificador da conta a ser removida",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Conta removida com sucesso (sem conteúdo)"
          },
          "400": {
            "description": "Requisição inválida (conta não pode ser deletada)"
          },
          "404": {
            "description": "Conta não encontrada"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Natureza": {
        "type": "string",
        "description": "Natureza da conta",
        "enum": [
          "CREDORA",
          "DEVEDORA"
        ]
      },
      "TipoConta": {
        "type": "string",
        "description": "Tipo da conta",
        "enum": [
          "ANALITICA",
          "SINTETICA"
        ]
      },
      "ContaNewDto": {
        "type": "object",
        "description": "DTO para criação de nova conta",
        "properties": {
          "superiorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "description": "ID da conta superior (obrigatório)"
          },
          "descricao": {
            "type": "string"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoConta"
          },
          "redutora": {
            "type": "boolean",
            "description": "Indica se a conta é redutora (opcional, default false)"
          }
        },
        "required": [
          "superiorId",
          "descricao",
          "tipo"
        ]
      },
      "ContaFlatDto": {
        "type": "object",
        "description": "Representação plana de uma conta",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "codigo": {
            "type": "string",
            "description": "Código calculado a partir do caminho (ex. \"1.2.3\")"
          },
          "descricao": {
            "type": "string"
          },
          "superiorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "redutora": {
            "type": "boolean",
            "description": "Indica se a conta é redutora (se a natureza é diferente da raiz)"
          },
          "saldoAtual": {
            "type": "number",
            "format": "double",
            "description": "Saldo atual da conta (pode ser nulo se não calculado)",
            "nullable": true
          },
          "deletable": {
            "type": "boolean",
            "description": "Indica se a conta pode ser deletada"
          },
          "editable": {
            "type": "boolean",
            "description": "Indica se a conta pode ser editada"
          },
          "natureza": {
            "$ref": "#/components/schemas/Natureza"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoConta"
          }
        },
        "required": [
          "id",
          "codigo",
          "descricao",
          "natureza",
          "tipo"
        ]
      },
      "ContaTreeDto": {
        "type": "object",
        "description": "Representação hierárquica de uma conta com seus filhos",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "codigo": {
            "type": "string",
            "description": "Código calculado a partir do caminho (ex. \"1.2.3\")"
          },
          "descricao": {
            "type": "string"
          },
          "natureza": {
            "$ref": "#/components/schemas/Natureza"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoConta"
          },
          "inferiores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContaTreeDto"
            },
            "description": "Contas filhas"
          },
          "redutora": {
            "type": "boolean",
            "description": "Indica se a conta é redutora (true quando a natureza for diferente da raiz)"
          },
          "saldoAtual": {
            "type": "number",
            "format": "double",
            "description": "Saldo atual da conta (pode ser nulo se não calculado)",
            "nullable": true
          },
          "editable": {
            "type": "boolean",
            "description": "Indica se a conta pode ser editada"
          },
          "deletable": {
            "type": "boolean",
            "description": "Indica se a conta pode ser deletada"
          }
        },
        "required": [
          "id",
          "codigo",
          "descricao",
          "natureza",
          "tipo"
        ]
      }
    }
  }
}