# src/main/java/me/josecomparotto/contabilidade_pessoal/model/dto/ContaTreeDto.java
package me.josecomparotto.contabilidade_pessoal.model.dto;

import java.math.BigDecimal;
import java.util.ArrayList;
import java.util.List;

import me.josecomparotto.contabilidade_pessoal.model.enums.Natureza;
import me.josecomparotto.contabilidade_pessoal.model.enums.TipoConta;

public class ContaTreeDto {
    private Integer id;
    private String codigo;
    private String descricao;
    private Natureza natureza;
    private TipoConta tipo;
    private Boolean redutora;
    private final List<ContaTreeDto> inferiores = new ArrayList<>();
    private Boolean editable;
    private Boolean deletable;

    // saldo atual (pode ser nulo se não calculado/populado)
    private BigDecimal saldoAtual;

    public ContaTreeDto() {
    }

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getCodigo() {
        return codigo;
    }

    public void setCodigo(String codigo) {
        this.codigo = codigo;
    }

    public String getDescricao() {
        return descricao;
    }

    public void setDescricao(String descricao) {
        this.descricao = descricao;
    }

    public List<ContaTreeDto> getInferiores() {
        return inferiores;
    }

    public Natureza getNatureza() {
        return natureza;
    }

    public void setNatureza(Natureza natureza) {
        this.natureza = natureza;
    }

    public TipoConta getTipo() {
        return tipo;
    }

    public void setTipo(TipoConta tipo) {
        this.tipo = tipo;
    }

    public boolean isRedutora() {
        return redutora;
    }

    public void setRedutora(boolean redutora) {
        this.redutora = redutora;
    }

    public BigDecimal getSaldoAtual() {
        return saldoAtual;
    }

    public void setSaldoAtual(BigDecimal saldoAtual) {
        this.saldoAtual = saldoAtual;
    }

    public Boolean isEditable() {
        return editable;
    }

    public void setEditable(Boolean editable) {
        this.editable = editable;
    }

    public Boolean isDeletable() {
        return deletable;
    }

    public void setDeletable(Boolean deletable) {
        this.deletable = deletable;
    }

    public Boolean getDeletable() {
        return deletable;
    }

}